--- Sources/Application/Routes/MultiHandlerRoutes.swift
@@ -1,30 +1,0 @@
-/**
- * Copyright IBM Corporation 2018
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- **/
-
-func initializeMultiHandlerRoutes(app: App) {
-    // Uses multiple handler blocks
-    app.router.get("/multi", handler: { request, response, next in
-        response.send("I'm here!\n")
-        next()
-    }, { request, response, next in
-        response.send("Me too!\n")
-        next()
-    })
-
-    app.router.get("/multi") { request, response, next in
-        try response.send("I come afterward..\n").end()
-    }
-}
