+++ Views/basicauth.html
@@ -79,7 +79,6 @@
      This page will send HTTP requests to the "/basic" and "/rawbasic" routes of the Kitura sample server.<br>
      These pages have been protected by <a href="https://developer.mozilla.org/en-US/docs/Web/HTTP/Authentication#Basic_authentication_scheme" target="_blank"> HTTP basic authentication</a>.<br>
      To access the routes you must sign up and then provide the correct username and password.<br>
-     The sample has a default username of "username" and password of "password".<br>
      Once you have successfully authenticated your browser will remember your credentials for future logins.<br><br>
      <a href="https://github.com/IBM-Swift/Kitura-Sample/blob/master/Sources/Application/Routes/BasicAuthRoutes.swift" target="_blank" class="inLink">View the code</a>
      </p>
@@ -95,11 +94,11 @@
        </div>
      </li></ul>
      <ul><li>
-         <p> "Codable HTTP Basic authentication" </p>
+         <p> Codable HTTP Basic authentication </p>
          <input type="button" value="GET to &quot;/basic&quot;" id="getButton" >
      </li></ul>
      <ul><li>
-         <p> "Raw HTTP Basic authentication" </p>
+         <p> Raw HTTP Basic authentication </p>
          <input type="button" value="GET to &quot;/rawbasic&quot;" id="rawGetButton" >
      </li></ul>
      <ul><li>
