+++ Views/further.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html>
+ <head>
+   <meta charset="UTF-8">
+   <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+   <link rel="stylesheet" type="text/css" href="stylesheet.css"/>
+   <title>Further routing examples</title>
+ </head>
+ <body>
+   <h2>Further routing examples</h2>
+     <p>
+         This page will send HTTP requests to demonstate additional features of Raw routing.
+         They will show the user of multiple handlers, redirecting, sending errors and custom path parameters.<br><br>
+         <a href="https://github.com/IBM-Swift/Kitura-Sample/blob/master/Sources/Application/Routes/FurtherRoutes.swift" target="_blank" class="inLink">View the code</a>
+     </p>
+     <br>
+     <ul><li>
+         <p>Send an HTTP GET request with multiple handlers:</p>
+         <input type="button" value="GET to &quot;/multi&quot;" onclick="window.location.href='/multi'">
+     </li></ul>
+     <ul><li>
+         <p>Send an HTTP GET request to a redirecting handler:</p>
+         <input type="button" value="GET to &quot;/redirect&quot;" onclick="window.location.href='/redirect'">
+     </li></ul>
+     <ul><li>
+         <p>Send an HTTP GET request to a handler that parses the URL path as a parameter:</p>
+         <input type="button" value="GET to &quot;/users/example&quot;" onclick="window.location.href='/users/example'">
+     </li></ul>
+     <ul><li>
+         <p>Send an HTTP GET request to a handler that returns an error:</p>
+         <input type="button" value="GET to &quot;/error&quot;" onclick="window.location.href='/error'">
+     </li></ul>
+ </body>
+</html>
