+++ Views/index.html
@@ -24,21 +24,19 @@
 
             <li class="list"><button class="button" onclick="loadInnerPage('./codable.html')"><p>Codable routing</p></button></li>
 
-            <li class="list"><button class="button" onclick="loadInnerPage('./additional.html')"><p>Further routing examples</p></button></li>
-
             <li class="list"><button class="button" onclick="loadInnerPage('./ormdatabase.html')"><p>Persisting with a database</p></button></li>
 
             <li class="list"><button class="button" onclick="loadInnerPage('./sessions.html')"><p>Persisting with sessions</p></button></li>
 
+            <li class="list"><button class="button" onclick="loadInnerPage('./articles')"><p>Rendering stencil templates</p></button></li>
+
+            <li class="list"><button class="button" onclick="loadInnerPage('./docs/index')"><p>Rendering markdown templates</p></button></li>
+            
             <li class="list"><button class="button" onclick="loadInnerPage('./basicauth.html')"><p>HTTP basic authentication</p></button></li>
-
+            
             <li class="list"><button class="button" onclick="loadInnerPage('./tokenauth.html')"><p>OAuth token authentication</p></button></li>
-
-            <li class="list"><button class="button" onclick="loadInnerPage('./articles')"><p>Rendering stencil template</p></button></li>
-
-            <li class="list"><button class="button" onclick="loadInnerPage('./docs/index')"><p>Rendering markdown template</p></button></li>
-
-            <li class="list"><button class="button" onclick="loadInnerPage('./error')"><p>Error messages</p></button></li>
+            
+            <li class="list"><button class="button" onclick="loadInnerPage('./further.html')"><p>Further routing examples</p></button></li>
 
             <li class="list"><button class="button" onclick="loadInnerPage('./chat')"><p>WebSockets demo</p></button></li>
 
@@ -52,22 +50,20 @@
             <li style="margin: 5px -15px"><button class="button" onclick="window.open('./hello.html')">./hello<p>Simple Hello World response in a new browser tab.</p></button></li>
 
             <li class="list"><button class="button" onclick="window.open('./codable.html')">./books<p>Codable route demo of fetching a struct object book.</p></button></li>
-
-            <li class="list"><button class="button" onclick="window.open('./additional.html')">./additional<p>Additional raw routing handlers.</p></button></li>
-
+            
             <li class="list"><button class="button" onclick="window.open('./ormdatabase.html')">./grades<p>Database example using the ORM.</p></button></li>
 
             <li class="list"><button class="button" onclick="window.open('./sessions.html')">./sessions<p>Codable and Raw sessions using cookies.</p></button></li>
 
-            <li class="list"><button class="button" onclick="window.open('./basicauth.html')">./basicauth<p>Codable and Raw routes using HTTP basic authentication .</p></button></li>
-
-            <li class="list"><button class="button" onclick="window.open('./tokenauth.html')">./tokenauth<p>Codable and Raw routes using OAuth token authentication .</p></button></li>
-
             <li class="list"><button class="button" onclick="window.open('./articles')">./articles<p>Render a page using a Stencil file and Swift code.</p></button></li>
 
             <li class="list"><button class="button" onclick="window.open('./docs/index')">./docs<p>Render a static page from a Markdown file using Markdown.</p></button></li>
-
-            <li class="list"><button class="button" onclick="window.open('./error')">./error<p>Displays an error message in a new tab.</p></button></li>
+            
+            <li class="list"><button class="button" onclick="window.open('./basicauth.html')">./basicauth<p>Codable and Raw routes using HTTP basic authentication .</p></button></li>
+            
+            <li class="list"><button class="button" onclick="window.open('./tokenauth.html')">./tokenauth<p>Codable and Raw routes using OAuth token authentication .</p></button></li>
+            
+            <li class="list"><button class="button" onclick="window.open('./further.html')"><p>Further routing examples</p></button></li>
 
             <li class="list"><button class="button" onclick="window.open('./chat')">./chat<p>Kitura WebSockets demo using a Chat Room application.</p></button></li>
