+++ Views/tokenauth.html
@@ -101,8 +101,8 @@
         /<a href="https://developers.google.com/identity/sign-in/ios/backend-auth" target="_blank">Google</a>
         in your mobile application. <br>
       </p>
-      <p>Create a google user access token using the <a href="https://developers.google.com/oauthplayground/" target="_blank">google oauth playground</a>.<br>
-         Create a facebook user access token using the <a href="https://developers.facebook.com/tools/explorer" target="_blank">facebook tools explorer</a>.</p>
+      <p>Create a google user access token using the <a href="https://developers.google.com/oauthplayground/" target="_blank">Google oauth playground</a>.<br>
+         Create a facebook user access token using the <a href="https://developers.facebook.com/tools/explorer" target="_blank">Facebook tools explorer</a>.</p>
       <h2>
           Access the protected routes
       </h2>
